@@ -110,42 +94,57 @@
     "koa-proxy": "^0.6.0",
     "koa-static": "^3.0.0",
     "material-ui": "^0.16.4",
-    "moment": "^2.18.1",
-    "node-fetch": "^1.6.3",
+    "moment": "^2.22.0",
+    "node-fetch": "^1.7.2",
     "node-sass": "^3.7.0",
     "normalize.css": "^4.1.1",
     "postcss-loader": "^0.9.0",
-    "react": "^15.4.1",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
     "react-bootstrap-toggle": "^2.0.4",
     "react-color": "^2.11.7",
     "react-date-picker": "^5.3.28",
     "react-date-range": "^0.2.4",
-    "react-dom": "^15.4.1",
+    "react-dom": "^15.6.1",
     "react-file-uploader": "^0.3.3",
     "react-google-login": "^2.3.2",
     "react-google-login-component": "^0.5.2",
+    "react-images": "^0.5.17",
+    "react-jsx-highcharts": "^1.3.4",
+    "react-jsx-highstock": "^1.3.4",
+    "react-month-picker": "^1.3.0",
     "react-redux": "^4.0.0",
     "react-router": "^2.2.0",
     "react-router-redux": "^4.0.0",
     "react-rte": "^0.11.0",
+    "react-signature": "^1.2.4",
+    "react-signature-canvas": "^0.3.0",
     "react-tap-event-plugin": "^2.0.1",
     "react-toastr": "^2.8.0",
     "react-toggle-button": "^2.1.0",
     "redux": "^3.0.0",
     "redux-actions": "^0.10.1",
     "redux-logger": "^2.6.1",
+    "redux-saga": "^0.15.4",
     "redux-thunk": "^2.0.0",
     "rimraf": "^2.5.1",
     "sass-loader": "^4.0.0",
+    "signature_pad": "^2.3.2",
     "style-loader": "^0.13.0",
+    "sweetalert": "^1.1.3",
+    "tether": "^1.4.0",
+    "timepicker": "^1.10.1",
     "toastr": "^2.1.2",
+    "underscore": "^1.8.3",
     "upload-file": "^1.2.0",
     "url-loader": "^0.5.6",
-    "webpack": "^1.12.14",
+    "webpack": "2.6.1",
     "whatwg-fetch": "^2.0.2",
     "yargs": "^4.0.0"
   },
   "devDependencies": {
+    "@types/signature_pad": "0.0.30",
+    "babel-core": "^6.25.0",
     "babel-eslint": "^6.0.0-beta.6",
     "chai": "^3.4.1",
     "chai-as-promised": "^5.1.0",
