@@ -1,74 +1,57 @@
-import React from 'react'
+import React from 'react';
+import PropTypes from 'prop-types';
 
-class UserHorizontalView extends React.Component {
-  constructor (props) {
-    super(props)
+const UserHorizontalView = ({profileImage, name, jobtitle, dateofjoining, gender, dob, workEmail, inventory}) => {
+  let joining, details;
+  if (!inventory) {
+    joining = <h5 className="text-muted">Joining Date: <b>{dateofjoining}</b></h5>;
+    details = <div className="col-xs-6 profile-input text-muted">
+      <h5>Gender: <b>{gender}</b></h5>
+      <h5>Date Of Birth: <b>{dob}</b></h5>
+      <h5>Work Email: <b>{workEmail}</b></h5>
+    </div>;
   }
-  componentWillReceiveProps (props) {}
-
-  render () {
-    return (
-
-      <div className="item">
-        <div className="item-bg">
-          <img src={this.props.profileImage} className="blur opacity-3" />
-        </div>
-        <div className="p-a-md">
-          <div className="row m-t">
-            <div className="col-sm-7">
-              <a href="" className="pull-left m-r-md">
-                <span className="avatar w-96">
-                  <img src={this.props.profileImage} />
-                  <i className="on b-white"></i>
-                </span>
-              </a>
-              <div className="clear m-b">
-                <h3 className="m-a-0 m-b-xs">{this.props.name}</h3>
-                <p className="text-muted">
-                  <span className="m-r">{this.props.jobtitle}</span>
-                </p>
-                {this.props.inventory ? null
-                  : <p className="text-muted">
-                    <span className="m-r">Joining Date : {this.props.dateofjoining}</span>
-                  </p>
-                }
-
-              </div>
-            </div>
-
-            {this.props.inventory ? null
-              : <div className="col-sm-5">
-                <p className="text-muted">
-                  <span className="m-r">Gender :
-                    <b>{this.props.gender}</b>
-                  </span>
-                </p>
-                <p className="text-muted">
-                  <span className="m-r">Date Of Birth :
-                    <b>{this.props.dob}</b>
-                  </span>
-                </p>
-                <p className="text-muted">
-                  <span className="m-r">Work Email :
-                    <b>{this.props.work_email}</b>
-                  </span>
-                </p>
-              </div>
-            }
-
-          </div>
-        </div>
+  return (
+    <div className="item">
+      <div className="item-bg">
+        <img src={profileImage} className="blur opacity-3" />
       </div>
+      <div className="row padding-responsive">
+        <div className="col-xs-12 m-b">
+          <span className="avatar w-96 pointer">
+            <img src={profileImage} />
+            <i className="on b-white"></i>
+          </span>
+        </div>
+        <div className="col-xs-6 profile-input">
+          <h3 className="text-capitalize m-t-xs">{name}</h3>
+          <h5 className="text-muted text-capitalize"><b>{jobtitle}</b></h5>
+          {joining}
+        </div>
+        {details}
+      </div>
+    </div>
+  );
+};
 
-    )
-  }
-}
 UserHorizontalView.propTypes = {
-  name: React.PropTypes.string,
-  jobtitle: React.PropTypes.string,
-  dateofjoining: React.PropTypes.string,
-  dob: React.PropTypes.string,
-  Gender: React.PropTypes.string,
-  Email: React.PropTypes.string
-}
-export default UserHorizontalView
+  profileImage:  PropTypes.string,
+  name:          PropTypes.string,
+  jobtitle:      PropTypes.string,
+  dateofjoining: PropTypes.string,
+  dob:           PropTypes.string,
+  gender:        PropTypes.string,
+  workEmail:     PropTypes.string,
+  inventory:     PropTypes.bool
+};
+UserHorizontalView.defaultProps = {
+  profileImage:  null,
+  name:          null,
+  jobtitle:      null,
+  dateofjoining: null,
+  dob:           null,
+  gender:        null,
+  workEmail:     null,
+  inventory:     false
+};
+export default UserHorizontalView;
