@@ -1,12 +1,15 @@
+import config from '../config'
 import Koa from 'koa'
 import convert from 'koa-convert'
 import webpack from 'webpack'
-import webpackConfig from '../build/webpack.config'
+import webpackConfigDevelopment from '../build/webpack.config.development'
+import webpackConfigProduction from '../build/webpack.config.production'
+
 import historyApiFallback from 'koa-connect-history-api-fallback'
 import serve from 'koa-static'
 import proxy from 'koa-proxy'
 import _debug from 'debug'
-import config from '../config'
+
 import webpackDevMiddleware from './middleware/webpack-dev'
 import webpackHMRMiddleware from './middleware/webpack-hmr'
 
@@ -26,23 +29,26 @@ app.use(convert(historyApiFallback({
   verbose: false
 })))
 
+let webpackConfig = webpackConfigDevelopment;
+if( config.env === 'production' ){  
+  webpackConfig = webpackConfigProduction;
+}
+
 // ------------------------------------
 // Apply Webpack HMR Middleware
 // ------------------------------------
-if (config.env === 'development') {
+if (config.env === 'development' || config.env === 'production')   {
   const compiler = webpack(webpackConfig)
-
   // Enable webpack-dev and webpack-hot middleware
   const { publicPath } = webpackConfig.output
-
   app.use(webpackDevMiddleware(compiler, publicPath))
   app.use(webpackHMRMiddleware(compiler))
-
   // Serve static assets from ~/src/static since Webpack is unaware of
   // these files. This middleware doesn't need to be enabled outside
   // of development since this directory will be copied into ~/dist
   // when the application is compiled.
-  app.use(serve(paths.client('static')))
+  //app.use(serve(paths.client('static')))
+  app.use(serve('./src/static'))
 } else {
   debug(
     'Server is being run outside of live development mode, meaning it will ' +
@@ -51,7 +57,6 @@ if (config.env === 'development') {
     'server such as nginx to serve your static files. See the "deployment" ' +
     'section in the README for more information on deployment strategies.'
   )
-
   // Serving ~/dist by default. Ideally these files should be served by
   // the web server and not the app server, but this helps to demo the
   // server in production.
