@@ -1,22 +1,22 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import DayReference from './DayReference';
+import TimeSummary from './TimeSummary';
+import CompensationSummary from './compensationSummary';
 
-import DayReference from './DayReference'
-import TimeSummary from './TimeSummary'
+const MonthSummary = ({monthlyAttendance}) => {
+  return (
+    <div>
+      <TimeSummary monthlyAttendance={monthlyAttendance} />
+      {monthlyAttendance.compensationSummary && monthlyAttendance.compensationSummary.seconds_to_be_compensate > 0 ? <CompensationSummary monthlyAttendance={monthlyAttendance} /> : ''}
+      <DayReference monthlyAttendance={monthlyAttendance} />
+    </div>
+  );
+};
 
-class MonthSummary extends React.Component {
-    constructor( props ){
-		super( props );
-    }
-    render(){
-      return (
-        <div>
-          <TimeSummary  {...this.props}/>
-          <DayReference  {...this.props}/>
-        </div>
-	    )
-    }
-}
+MonthSummary.PropTypes = {
+  monthlyAttendance:   PropTypes.object.isRequired,
+  compensationSummary: PropTypes.object.isRequired
+};
 
-export default MonthSummary
-
-
+export default MonthSummary;
