@@ -1,66 +1,69 @@
-// import React from 'react'
-// import ReactDOM from 'react-dom'
-// import createBrowserHistory from 'history/lib/createBrowserHistory'
-// import { useRouterHistory } from 'react-router'
-// import { syncHistoryWithStore } from 'react-router-redux'
-// import createStore from './store/createStore'
-// import AppContainer from './containers/AppContainer'
+import 'styles/main.scss';
+import 'whatwg-fetch';
+import 'sweetalert';
+import 'sweetalert/dist/sweetalert.css';
+import 'jquery';
+import 'jquery-ui';
+import 'tether';
+import 'bootstrap';
+import './themeFlatkit/scripts';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import {Router, Route, IndexRoute, useRouterHistory} from 'react-router';
+import {createHashHistory} from 'history';
+import {Provider} from 'react-redux';
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import injectTapEventPlugin from 'react-tap-event-plugin';
 
-// --start---for HR APP by arun
-import React from 'react'
-import ReactDOM from 'react-dom'
-import {Router, Route, Link, IndexRoute, hashHistory} from 'react-router'
-import {createStore, applyMiddleware, compose} from 'redux'
-import {Provider} from 'react-redux'
-import Immutable from 'immutable'
-import thunk from 'redux-thunk'
-import createLogger from 'redux-logger'
+import Page_Login from './modules/auth/containers/login';
+import Page_Logout from './modules/auth/containers/logout';
+import Page_ForgotPassword from './modules/auth/containers/forgotPassword';
+import Page_ManageLeaves from './modules/leave/containers/manageLeaves';
+import Page_LeavesSummary from './modules/leave/containers/leavesSummary';
+import Page_ApplyLeave from './modules/leave/containers/applyLeave';
+import Page_MyLeaves from './modules/leave/containers/myLeaves';
+import Page_Holidays from './modules/holidays/containers/holidays';
+import Page_PolicyDocument from './modules/policyDocuments/containers/PolicyDocument';
+import Page_UploadPolicyDocument from './modules/policyDocuments/containers/uploadPolicyDocument';
+import Page_ManageSalary from './modules/salary/containers/manageSalary';
+import Page_ViewSalary from './modules/salary/containers/viewSalary';
+import Page_Salary from './modules/salary/containers/salary';
+import Page_ManagePayslips from './modules/salary/containers/managePayslips';
+import Page_Home from './modules/attendance/containers/Home';
+import Page_MonthlyAttendance from './modules/attendance/containers/monthlyAttendance';
+import Page_UploadAttendance from './modules/attendance/containers/uploadAttendance';
+import Page_MyDocuments from './modules/myDocuments/containers/myDocuments';
+import PageManageUsers from './modules/manageUsers/containers/manageUsers';
+import Page_ManageWorkingHours from './modules/workingHours/containers/manageWorkingHours';
+import Page_ManageUserWorkingHours from './modules/workingHours/containers/manageUserWorkingHours';
+import InventorySystem from './modules/inventory/containers/manageInventory';
+import Page_MyInventory from './modules/inventory/containers/myInventory';
+import PageManageRoles from './modules/manageRoles/containers/manageRoles';
+import Page_ManageUserPendingHours from './modules/workingHours/containers/manageUserPendingHours';
+import PageDisabledEmployes from 'modules/manageUsers/containers/disabledEmployes';
+import Page_mail_template from './modules/templates/containers/addTemplate';
+import Page_AddVariables from './modules/templates/containers/addVariables';
+import Page_TeamView from './modules/team/containers/viewTeam';
+import Page_MyProfile from './modules/myProfile/containers/myProfile';
+import PageManageDashboard from 'modules/manageUsers/containers/manageDashboard';
+import Page_AttendanceReq from './modules/attendance/containers/manageAttendanceReq';
+import InventoryItem from './modules/inventory/components/inventoryItem'
+import Page_AddDocuments from './modules/addDocuments/container/AddDocuments';
+import Page_FormAddDocuments from './modules/addDocuments/components/FormAddDocuments';
+import FormAddNewEmployeeDetails from './modules/manageUsers/components/FormAddNewEmployeeDetails'
+import AddNewEmployee from './modules/manageUsers/components/FormAddNewEmployee'
+import UserDocumentDetails from './modules/manageUsers/components/Userdocuments'
+import store from './store';
+import Page_InventoryOverview from './modules/inventoryOverview/container/inventoryOverviewContainer.jsx';  
+import RouterAddInventorySystem from './modules/inventory/components/AddInventory'
+// import APP from './App';
 
-import reducer from './reducers/index'
-
-import injectTapEventPlugin from 'react-tap-event-plugin'
-
-// pages
-import Page_Login from './containers/user/login'
-import Page_Home from './containers/user/home'
-import Page_Logout from './containers/user/logout'
-import Page_ForgotPassword from './containers/user/forgotPassword'
-import Page_PolicyDocument from './containers/user/PolicyDocument'
-
-// -admin
-import Page_ManageWorkingHours from './containers/admin/manageWorkingHours'
-import Page_ManageLeaves from './containers/admin/manageLeaves'
-import Page_ManageUserWorkingHours from './containers/admin/manageUserWorkingHours'
-import Page_LeavesSummary from './containers/admin/leavesSummary'
-import Page_ManageSalary from './containers/admin/manageSalary'
-import Page_ManageUsers from './containers/admin/manageUsers'
-import Page_ManageClients from './containers/admin/manageClients'
-import Page_ManagePayslips from './containers/admin/managePayslips'
-import Page_ViewSalary from './containers/admin/viewSalary'
-import Page_DisabledEmployes from './containers/admin/disabledEmployes'
-import Page_UploadAttendance from './containers/admin/uploadAttendance'
-import Page_UploadPolicyDocument from './containers/admin/uploadPolicyDocument'
-import Page_mail_template from './containers/admin/addTemplate'
-import Page_AddVariables from './containers/admin/addVariables'
-import Page_TeamView from './containers/admin/viewTeam'
-import Page_InventorySystem from './containers/admin/manageInventory'
-
-// -user
-import Page_MonthlyAttendance from './containers/user/monthlyAttendance'
-import Page_AttendanceSummary from './containers/user/attendanceSummary'
-import Page_Holidays from './containers/user/holidays'
-import Page_ApplyLeave from './containers/user/applyLeave'
-import Page_MyLeaves from './containers/user/myLeaves'
-import Page_Salary from './containers/user/salary'
-import Page_MyProfile from './containers/user/myProfile'
-import Page_MyInventory from './containers/user/myInventory'
-import Page_MyDocuments from './containers/user/myDocuments'
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
-import 'whatwg-fetch'
 // -for iPhone iPad safari engine
 if (!!navigator.platform && /iPad|iPhone|iPod/.test(navigator.platform)) {
-  fetch = require('whatwg-fetch')
+  fetch = require('whatwg-fetch');
 }
+const appHistory = useRouterHistory(createHashHistory)({queryKey: false});
+injectTapEventPlugin();
 
 export class APP extends React.Component {
   render () {
@@ -68,80 +71,21 @@ export class APP extends React.Component {
       <div>
         {this.props.children}
       </div>
-    )
+    );
   }
 }
 
-const logger = createLogger()
-
-injectTapEventPlugin()
-
-let store = createStore(reducer, compose(
-// applyMiddleware( thunk, logger),
-applyMiddleware(thunk),
-//window.devToolsExtension ? window.devToolsExtension() : f => f
-))
-
-// --end---for HR APP by arun
-
-// ========================================================
-// Browser History Setup
-// ========================================================
-// const browserHistory = useRouterHistory(createBrowserHistory)({
-//   basename: __BASENAME__
-// })
-
-// ========================================================
-// Store and History Instantiation
-// ========================================================
-// Create redux store and sync with react-router-redux. We have installed the
-// react-router-redux reducer under the routerKey "router" in src/routes/index.js,
-// so we need to provide a custom `selectLocationState` to inform
-// react-router-redux of its location.
-// const initialState = window.___INITIAL_STATE__
-// const store = createStore(initialState, browserHistory)
-// const history = syncHistoryWithStore(browserHistory, store, {
-//   selectLocationState: (state) => state.router
-// })
-
-// ========================================================
-// Developer Tools Setup
-// ========================================================
-// if (__DEBUG__) {
-//   if (window.devToolsExtension) {
-//     window.devToolsExtension.open()
-//   }
-// }
-
-// ========================================================
-// Render Setup
-// ========================================================
-// const MOUNT_NODE = document.getElementById('root')
-
-// let render = (routerKey = null) => {
-//   const routes = require('./routes/index').default(store)
-
-//   ReactDOM.render(
-//     <AppContainer
-//       store={store}
-//       history={history}
-//       routes={routes}
-//       routerKey={routerKey}
-//     />,
-//     MOUNT_NODE
-//   )
-// }
-
-// --start----added by arun for HR app
 let render = (routerKey = null) => {
   ReactDOM.render((
     <MuiThemeProvider>
       <Provider store={store}>
-        <Router history={hashHistory}>
+        <Router history={appHistory} queryKey={false}>
           <Route path="/" component={APP}>
-            <IndexRoute component={Page_Login} />
+            <IndexRoute component={PageManageDashboard} />
             //this will be the default page which will opens when app starts
             <Route path="home" component={Page_Home} />
+            <Route path="PageManageDashboard" component={PageManageDashboard} />
+            <Route path="page_login" component={Page_Login} />
             <Route path="monthly_attendance" component={Page_MonthlyAttendance} />
             <Route path="manage_working_hours" component={Page_ManageWorkingHours} />
             <Route path="logout" component={Page_Logout} />
@@ -150,15 +94,16 @@ let render = (routerKey = null) => {
             <Route path="apply_leave" component={Page_ApplyLeave} />
             <Route path="manage_leaves" component={Page_ManageLeaves} />
             <Route path="my_leaves" component={Page_MyLeaves} />
-            <Route path="disabled_employes" component={Page_DisabledEmployes} />
+            <Route path="disabled_employes" component={PageDisabledEmployes} />
             <Route path="manage_user_working_hours" component={Page_ManageUserWorkingHours} />
+            <Route path="manage_user_pending_hours" component={Page_ManageUserPendingHours} />
             <Route path="leaves_summary" component={Page_LeavesSummary} />
             <Route path="salary" component={Page_Salary} />
             <Route path="manage_salary" component={Page_ManageSalary} />
             <Route path="my_profile" component={Page_MyProfile} />
             <Route path="my_inventory" component={Page_MyInventory} />
-            <Route path="manage_users" component={Page_ManageUsers} />
-            <Route path="manage_clients" component={Page_ManageClients} />
+            <Route path="manage_users" component={PageManageUsers} />
+            <Route path="manage_roles" component={PageManageRoles} />
             <Route path="manage_payslips" component={Page_ManagePayslips} />
             <Route path="forgot_password" component={Page_ForgotPassword} />
             <Route path="documents" component={Page_MyDocuments} />
@@ -168,35 +113,27 @@ let render = (routerKey = null) => {
             <Route path="upload_policy_documents" component={Page_UploadPolicyDocument} />
             <Route path="add_variables" component={Page_AddVariables} />
             <Route path="mail_templates" component={Page_mail_template} />
-            <Route path="inventory_system" component={Page_InventorySystem} />
+
+            {/* <Route path="inventoryOverviewDetail" component={Page_InventorySystem} /> */}
+            {/* <Route path="inventory_system/:device/:id" component={InventoryItem} /> */}
+            <Route path="inventory_system" component={InventorySystem} />
+
+            <Route path="inventory_system/:device/:id" component={InventoryItem} />
+            <Route path="inventory_system/:device" component={InventorySystem}/>
+            <Route path="attendanceReq" component={Page_AttendanceReq} />
+            <Route path="inventoryOverviewDetail" component={Page_InventoryOverview}/>
+            <Route path="add_documents" component = {Page_AddDocuments} />
+            <Route path="add_documents/*" component = {Page_FormAddDocuments} />
+
+            <Route path="addInventory" component={RouterAddInventorySystem}/>
+            <Route path="add_new_employee" component={AddNewEmployee}/>
+            <Route path="add_new_employee/:id" component={FormAddNewEmployeeDetails}/>
+            <Route path="user_document" component={UserDocumentDetails}/>
           </Route>
         </Router>
       </Provider>
     </MuiThemeProvider>
-  ), document.querySelector('#myApp'))
-}
-// --end------added by arun for HR app
+  ), document.querySelector('#myApp'));
+};
 
-// Enable HMR and catch runtime errors in RedBox
-// This code is excluded from production bundle
-// if (__DEV__ && module.hot) {
-//   const renderApp = render
-//   const renderError = (error) => {
-//     const RedBox = require('redbox-react').default
-
-//     ReactDOM.render(<RedBox error={error} />, MOUNT_NODE)
-//   }
-//   render = () => {
-//     try {
-//       renderApp(Math.random())
-//     } catch (error) {
-//       renderError(error)
-//     }
-//   }
-//   //module.hot.accept(['./routes/index'], () => render())
-// }
-
-// ========================================================
-// Go!
-// ========================================================
-render()
+render();
