@@ -1,116 +1,109 @@
-import {CONFIG} from '../config/index'
-import * as _ from 'lodash'
-import 'whatwg-fetch'
-export function notify (text) {
-  alert(text)
-}
+import {CONFIG} from 'src/config/index';
+import _ from 'lodash';
+import {confirm} from 'src/services/notify';
+import {getToken, resetLoggedUser} from 'src/services/generic';
+import 'whatwg-fetch';
+import axios from 'axios';
 
-export function fireAjax (method, url, data) {
-  let URL = CONFIG.api_url + url
-  let headers = {}
-  if (method == 'GET') {
-    headers = {
-      method: 'GET',
-      mode: 'cors',
-      cache: 'no-cache'
-    }
-  } else if (method == 'POST') {
-    let token = localStorage.getItem('hr_logged_user')
-    data.token = token
-    headers = {
-      method: 'POST',
-      mode: 'cors',
-      cache: 'no-cache',
-      body: JSON.stringify(data)
-    }
-  }
+const actionsForOtherAPIurl = ['get_user_profile_detail', 'get_user_profile_detail_by_id', 'update_user_profile_detail_by_id', 'update_user_bank_detail',
+  'update_user_profile_detail', 'get_user_manage_payslips_data', 'create_employee_salary_slip', 'delete_salary',
+  'send_payslips_to_employees', 'get_user_document','get_user_document_by_id', 'insert_user_document', 'delete_user_document', 'get_all_users_detail',
+  'create_template_variable', 'get_template_variable', 'delete_template_variable', 'update_template_variable', 'create_email_template',
+  'get_email_template', 'delete_email_template', 'update_email_template', 'send_employee_email', 'create_pdf', 'get_policy_document',
+  'save_policy_document', 'get_user_policy_document', 'update_user_policy_document', 'add_team_list', 'get_team_list',
 
-  if (data.action == 'get_salary_details') {
-    let token = localStorage.getItem('hr_logged_user')
-    URL = CONFIG.api_url_salary + '/salary_info.php?token=' + token
-  } else if (data.action == 'get_user_salary_details') {
-    let token = localStorage.getItem('hr_logged_user')
-    URL = CONFIG.api_url_salary + '/salary_info.php?token=' + token + '&user_id=' + data.userid
-    console.log('api URL for get_user_salary_details: ', URL)
-  } else if (data.action == 'add_user_salary') {
-    let token = localStorage.getItem('hr_logged_user')
-    delete (data.action)
-    data.token = token
-    headers = {
-      method: 'POST',
-      mode: 'cors',
-      cache: 'no-cache',
-      body: JSON.stringify(data)
-    }
-    URL = CONFIG.api_url_salary + '/add_sal_structure.php'
-  } else if (data.action == 'add_user_holding') {
-    let token = localStorage.getItem('hr_logged_user')
-    delete (data.action)
-    data.token = token
-    headers = {
-      method: 'POST',
-      mode: 'cors',
-      cache: 'no-cache',
-      body: JSON.stringify(data)
-    }
-    URL = CONFIG.api_url_salary + '/add_holding_info.php'
-  }
-  // else if( data.action == "update_user_bank_details" ){
-  // 	let token = localStorage.getItem('hr_logged_user')
-  // 	delete( data.action )
-  // 	data.token = token
-  // 	headers = {
-  // 		method : 'POST',
-  // 		mode: 'cors',
-  // 		cache: 'no-cache',
-  // 		body: JSON.stringify(data),
-  // 	}
-  // 	URL = CONFIG.api_url_salary +'/user_bank_detail.php'}
-  else if (data.action == 'get_user_profile_detail' || data.action == 'update_user_bank_detail' || data.action == 'update_user_profile_detail' ||
-    data.action == 'get_all_clients' || data.action == 'get_client_detail' || data.action == 'create_new_client' || data.action == 'create_client_invoice' ||
-    data.action == 'update_client_details' || data.action == 'delete_invoice' || data.action == 'get_user_manage_payslips_data' ||
-    data.action == 'create_employee_salary_slip' || data.action == 'delete_salary' || data.action == 'send_payslips_to_employees' ||
-    data.action == 'get_user_document' || data.action == 'insert_user_document' || data.action == 'delete_user_document' || data.action == 'get_all_users_detail' ||
-    data.action == 'create_template_variable' || data.action == 'get_template_variable' || data.action == 'delete_template_variable' ||
-    data.action == 'update_template_variable' || data.action == 'create_email_template' || data.action == 'get_email_template' || data.action == 'delete_email_template' ||
-    data.action == 'update_email_template' || data.action == 'send_employee_email' || data.action == 'create_pdf' || data.action == 'get_policy_document' ||
-    data.action == 'save_policy_document' || data.action == 'get_user_policy_document' || data.action == 'update_user_policy_document' ||
-    data.action == 'add_team_list' || data.action == 'get_team_list' || data.action == 'get_team_users_detail') { // generic other new api url
-    let token = localStorage.getItem('hr_logged_user')
-    data.token = token
-    headers = {
-      method: 'POST',
-      mode: 'cors',
-      cache: 'no-cache',
-      body: JSON.stringify(data)
-    }
-    URL = CONFIG.other_api_url
-  } else if (data.action == 'admin_user_apply_leave' || data.action == 'change_employee_status' || data.action == 'show_disabled_users') {
-    let token = localStorage.getItem('hr_logged_user')
-    data.token = token
-    headers = {
-      method: 'POST',
-      mode: 'cors',
-      cache: 'no-cache',
-      body: JSON.stringify(data)
-    }
-    URL = CONFIG.api_url
-  }
+  'get_team_users_detail', 'get_user_salary_info', 'get_user_salary_info_by_id', 'get_unassigned_machine_list' , 'add_user_comment' ];
 
+
+const actionsForAPIurl = ['admin_user_apply_leave','get_my_inventories', 'get_machine','change_employee_status', 'get_employee_life_cycle', 'update_employee_life_cycle', 'show_disabled_users', 'add_roles', 'list_all_roles', 'update_role', 'assign_user_role', 'delete_role', 'get_employee_monthly_hours', 'get_employee_performance'];
+
+const actionForExpressWeburl = ['update_time_by_employee', 'manual', 'approval'];
+
+export function fireAjax (method, url, data, api) {
+  let URL = CONFIG.api_url + url;
+  let action = data.action;
+  let token = getToken();
+  if (data.action !== 'get_team_stats' && data.action !== 'get_user_list' && data.action !== 'get_employee_hours' && data.action !== 'get_employee_performance' && data.action !== 'get_employee_monthly_hours' && data.action !== 'get_termination_joining_stats') {
+    data.token = token;
+  }
+  let headers = {
+    method: method,
+    mode:   'cors',
+    cache:  'no-cache',
+    Accept: 'application/json',
+    body:   JSON.stringify(data)
+  };
+  if (data.action === 'add_user_salary') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.api_url_salary + '/add_sal_structure.php';
+  } else if (data.action === 'add_user_holding') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.api_url_salary + '/add_holding_info.php';
+  } else if (_.indexOf(actionsForOtherAPIurl, data.action) >= 0) {
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.other_api_url;
+  } 
+  else if (_.indexOf(actionsForAPIurl, data.action) >= 0) {
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.api_url;
+  } else if (data.action === 'get_team_stats') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.expressApiUrl;
+  } else if (data.action === 'get_termination_joining_stats') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.expressApiUrl;
+  } else if (data.action === 'get_employee_hours') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.expressApiUrl;
+  } else if (data.action === 'manual') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.expressApiurl + '/attendance/manual';
+  } if (data.action === 'approval') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.expressApiUrl + '/attendance/approval';
+  } else if (data.action === 'update_time_by_employee') {
+    delete (data.action);
+    headers['Content-Type'] = 'application/x-www-form-urlencoded';
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.expressApiUrl + '/attendance/update_time_by_employee';
+  } else if (data.action === 'get_employee_monthly_hours') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.expressApiUrl;
+  } else if (data.action === 'get_employee_performance') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.expressApiUrl;
+  } else if (data.action === 'get_user_list') {
+    delete (data.action);
+    headers.body = JSON.stringify(data);
+    URL = CONFIG.expressApiUrl;
+  }
   return fetch(URL, headers).then((response) => {
     if (response.status === 500) {
       return new Promise((resolve, reject) => {
         response.json().then((data) => {
-          reject(data)
-        })
-      })
+          reject(data);
+        });
+      });
     } else if (response.status === 401) {
-      // alert('401 hai ')	;
-      localStorage.removeItem('hr_logged_user')
-      let login_page_url = CONFIG.login_page_url
-      location.href = login_page_url
+      confirm('401 Access Denied !', '<span style="color:#f27474;font-size:18px;font-weight:600">' + action + '</span><br/>You are unauthorized to the Action - Contact Admin!!', 'error').then((res) => {
+        resetLoggedUser();
+        location.href = CONFIG.BASE_URL;
+      });
     } else {
-      return response.json()
+      return response.json();
     }
-  })
+  });
+}
+
+export function uploadfile(formData, url) {  
+  return axios.post(url, formData);
 }
